feat(products): add role-based access control for product updates and deletions
Implement authorization checks to ensure only admins or product owners can update or delete products

+++ src/products/products.service.ts
@@ -1,10 +1,12 @@
 import {
+  ForbiddenException,
   Injectable,
   NotFoundException,
   UnprocessableEntityException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserEntity } from 'src/users/entities/user.entity';
+import { Role } from 'src/utility/enums/roles.enum';
 import { Repository } from 'typeorm';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
@@ -55,15 +57,26 @@ export class ProductsService {
     return product;
   }
 
-  async update(id: string, updateProductDto: UpdateProductDto) {
+  async update(
+    id: string,
+    updateProductDto: UpdateProductDto,
+    user: UserEntity,
+  ) {
     const existingProduct = await this.productRepository.findOne({
       where: {
         id,
       },
+      relations: ['user'],
     });
 
     if (!existingProduct) {
       throw new NotFoundException('Product not found');
+    }
+
+    if (user.role !== Role.ADMIN && user.id !== existingProduct.user.id) {
+      throw new ForbiddenException(
+        'You are not authorized to update this product',
+      );
     }
 
     await this.productRepository.update(id, updateProductDto);
@@ -76,7 +89,23 @@ export class ProductsService {
     return updatedProduct;
   }
 
-  remove(id: string) {
+  async remove(id: string, user: UserEntity) {
+    const existingProduct = await this.productRepository.findOne({
+      where: {
+        id,
+      },
+      relations: ['user'],
+    });
+
+    if (!existingProduct) {
+      throw new NotFoundException('Product not found');
+    }
+
+    if (user.role !== Role.ADMIN && user.id !== existingProduct.user.id) {
+      throw new ForbiddenException(
+        'You are not authorized to delete this product',
+      );
+    }
     return this.productRepository.delete(id);
   }
 }
